hot-fix: further fixed syntax MTEX

+++ models/CAM_based/MTEX/__init__.py
@@ -25,13 +25,13 @@ class MTEX(nn.Module):
         self.cnn_layers = nn.Sequential(
             OrderedDict(
                 [
-                    ("conv_1", nn.Conv2d(1, 16, (np.ceil(time_length / 2), 1))),
+                    ("conv_1", nn.Conv2d(1, 16, (int(np.ceil(time_length / 2)), 1))),
                     ("relu_1", nn.ReLU(inplace=True)),
-                    ("conv_2", nn.Conv2d(16, 32, (np.ceil(time_length / 4), 1))),
+                    ("conv_2", nn.Conv2d(16, 32, (int(np.ceil(time_length / 4)), 1))),
                     ("relu_2", nn.ReLU(inplace=True)),
                     ("conv_3", nn.Conv2d(32, 1, 1)),
                     ("relu_3", nn.ReLU(inplace=True)),
-                    ("view", View((feature_length, (np.ceil(time_length / 2))))),
+                    ("view", View((feature_length, (int(np.ceil(time_length / 2)))))),
                     ("conv_4", nn.Conv1d(feature_length, 64, 3)),
                     ("relu_4", nn.ReLU(inplace=True)),
                 ]
@@ -41,7 +41,7 @@ class MTEX(nn.Module):
         self.linear_layers = nn.Sequential(
             OrderedDict(
                 [
-                    ("fc1", nn.Linear(64 * np.ceil(time_length / 4), 32)),
+                    ("fc1", nn.Linear(64 * int(np.ceil(time_length / 4)), 32)),
                     ("fc2", nn.Linear(32, n_classes)),
                 ]
             )
